refactor(auth): split saml.go and platform.go by concern
saml.go (1292 → 4 files):
- saml.go: core service, assertion parsing, metadata, user extraction
- saml_session.go: SAML session management, assertion replay checks
- saml_logout.go: logout request/response generation and parsing

platform.go (1256 → 4 files):
- platform.go: DashboardAuthService core, Login, CreateUser, tokens
- platform_sso.go: SSO identity lookup, creation, linking
- platform_password.go: password reset, change, profile update, delete
- platform_mfa.go: TOTP setup, enable, verify, disable
