+++ internal/auth/platform.go
@@ -5,17 +5,14 @@ import (
 	"crypto/rand"
 	"crypto/sha256"
 	"encoding/base32"
-	"encoding/base64"
 	"encoding/hex"
 	"errors"
 	"fmt"
 	"net"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
-	"github.com/pquerna/otp/totp"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/crypto/bcrypt"
 
@@ -361,306 +358,6 @@ func (s *DashboardAuthService) Login(ctx context.Context, email, password string
 	}, nil
 }
 
-// ChangePassword changes a dashboard user's password
-func (s *DashboardAuthService) ChangePassword(ctx context.Context, userID uuid.UUID, currentPassword, newPassword string, ipAddress net.IP, userAgent string) error {
-	// Validate new password length
-	if len(newPassword) < MinPasswordLength {
-		return fmt.Errorf("password must be at least %d characters", MinPasswordLength)
-	}
-	if len(newPassword) > MaxPasswordLength {
-		return fmt.Errorf("password must be at most %d characters", MaxPasswordLength)
-	}
-
-	// Fetch current password hash
-	var currentHash string
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT password_hash FROM platform.users WHERE id = $1 AND deleted_at IS NULL
-		`, userID).Scan(&currentHash)
-	})
-	if err != nil {
-		return fmt.Errorf("failed to fetch user: %w", err)
-	}
-
-	// Verify current password
-	err = bcrypt.CompareHashAndPassword([]byte(currentHash), []byte(currentPassword))
-	if err != nil {
-		return errors.New("current password is incorrect")
-	}
-
-	// Hash new password
-	newHash, err := bcrypt.GenerateFromPassword([]byte(newPassword), DefaultBcryptCost)
-	if err != nil {
-		return fmt.Errorf("failed to hash password: %w", err)
-	}
-
-	// Update password
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET password_hash = $1, updated_at = NOW()
-			WHERE id = $2
-		`, newHash, userID)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to update password: %w", err)
-	}
-
-	// Log activity
-	s.logActivity(ctx, userID, "password_change", "user", userID.String(), ipAddress, userAgent, nil)
-
-	return nil
-}
-
-// UpdateProfile updates a dashboard user's profile information
-func (s *DashboardAuthService) UpdateProfile(ctx context.Context, userID uuid.UUID, fullName string, avatarURL *string) error {
-	// Validate full name
-	if err := ValidateName(fullName); err != nil {
-		return fmt.Errorf("invalid name: %w", err)
-	}
-
-	// Validate avatar URL if provided
-	if avatarURL != nil {
-		if err := ValidateAvatarURL(*avatarURL); err != nil {
-			return fmt.Errorf("invalid avatar URL: %w", err)
-		}
-	}
-
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET full_name = $1, avatar_url = $2, updated_at = NOW()
-			WHERE id = $3 AND deleted_at IS NULL
-		`, fullName, avatarURL, userID)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to update profile: %w", err)
-	}
-
-	return nil
-}
-
-// DeleteAccount soft-deletes a dashboard user account
-func (s *DashboardAuthService) DeleteAccount(ctx context.Context, userID uuid.UUID, password string, ipAddress net.IP, userAgent string) error {
-	// Verify password
-	var passwordHash string
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT password_hash FROM platform.users WHERE id = $1 AND deleted_at IS NULL
-		`, userID).Scan(&passwordHash)
-	})
-	if err != nil {
-		return fmt.Errorf("failed to fetch user: %w", err)
-	}
-
-	err = bcrypt.CompareHashAndPassword([]byte(passwordHash), []byte(password))
-	if err != nil {
-		return errors.New("password is incorrect")
-	}
-
-	// Soft delete account
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET deleted_at = NOW(), updated_at = NOW()
-			WHERE id = $1
-		`, userID)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to delete account: %w", err)
-	}
-
-	// Delete all sessions
-	_ = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			DELETE FROM platform.sessions WHERE user_id = $1
-		`, userID)
-		return err
-	})
-
-	// Log activity
-	s.logActivity(ctx, userID, "account_delete", "user", userID.String(), ipAddress, userAgent, nil)
-
-	return nil
-}
-
-// SetupTOTP generates a new TOTP secret for 2FA
-// If issuer is empty, uses the configured default
-func (s *DashboardAuthService) SetupTOTP(ctx context.Context, userID uuid.UUID, email string, issuer string) (string, string, error) {
-	// Use provided issuer, or fall back to configured default
-	if issuer == "" {
-		issuer = s.totpIssuer
-	}
-
-	// Generate TOTP secret with QR code as data URI
-	secret, qrCodeDataURI, _, err := GenerateTOTPSecret(issuer, email)
-	if err != nil {
-		return "", "", fmt.Errorf("failed to generate TOTP secret: %w", err)
-	}
-
-	// Store secret (not yet enabled)
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET totp_secret = $1, totp_enabled = false, updated_at = NOW()
-			WHERE id = $2
-		`, secret, userID)
-		return err
-	})
-	if err != nil {
-		return "", "", fmt.Errorf("failed to store TOTP secret: %w", err)
-	}
-
-	// Return secret and QR code data URI
-	return secret, qrCodeDataURI, nil
-}
-
-// EnableTOTP enables 2FA after verifying the TOTP code
-func (s *DashboardAuthService) EnableTOTP(ctx context.Context, userID uuid.UUID, code string, ipAddress net.IP, userAgent string) ([]string, error) {
-	// Fetch TOTP secret
-	var secret string
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT totp_secret FROM platform.users WHERE id = $1 AND deleted_at IS NULL
-		`, userID).Scan(&secret)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch TOTP secret: %w", err)
-	}
-
-	if secret == "" {
-		return nil, errors.New("TOTP not set up")
-	}
-
-	// Verify code
-	valid := totp.Validate(code, secret)
-	if !valid {
-		return nil, errors.New("invalid TOTP code")
-	}
-
-	// Generate backup codes
-	backupCodes := make([]string, 10)
-	hashedBackupCodes := make([]string, 10)
-	for i := 0; i < 10; i++ {
-		code, err := generateBackupCode()
-		if err != nil {
-			return nil, fmt.Errorf("failed to generate backup code: %w", err)
-		}
-		backupCodes[i] = code
-
-		// Hash the backup code
-		hash, err := bcrypt.GenerateFromPassword([]byte(code), DefaultBcryptCost)
-		if err != nil {
-			return nil, fmt.Errorf("failed to hash backup code: %w", err)
-		}
-		hashedBackupCodes[i] = string(hash)
-	}
-
-	// Enable TOTP and store backup codes
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET totp_enabled = true, backup_codes = $1, updated_at = NOW()
-			WHERE id = $2
-		`, hashedBackupCodes, userID)
-		return err
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to enable TOTP: %w", err)
-	}
-
-	// Log activity
-	s.logActivity(ctx, userID, "2fa_enable", "user", userID.String(), ipAddress, userAgent, nil)
-
-	return backupCodes, nil
-}
-
-// VerifyTOTP verifies a TOTP code during login
-func (s *DashboardAuthService) VerifyTOTP(ctx context.Context, userID uuid.UUID, code string) error {
-	// Fetch TOTP secret and backup codes
-	var secret string
-	var backupCodes []string
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT totp_secret, COALESCE(backup_codes, ARRAY[]::text[])
-			FROM platform.users
-			WHERE id = $1 AND deleted_at IS NULL AND totp_enabled = true
-		`, userID).Scan(&secret, &backupCodes)
-	})
-	if err != nil {
-		return fmt.Errorf("failed to fetch TOTP data: %w", err)
-	}
-
-	// Try TOTP code first
-	valid := totp.Validate(code, secret)
-	if valid {
-		return nil
-	}
-
-	// Try backup codes
-	for i, hashedCode := range backupCodes {
-		err := bcrypt.CompareHashAndPassword([]byte(hashedCode), []byte(code))
-		if err == nil {
-			// Remove used backup code
-			backupCodes = append(backupCodes[:i], backupCodes[i+1:]...)
-			err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-				_, err := tx.Exec(ctx, `
-					UPDATE platform.users
-					SET backup_codes = $1, updated_at = NOW()
-					WHERE id = $2
-				`, backupCodes, userID)
-				return err
-			})
-			if err != nil {
-				return fmt.Errorf("failed to update backup codes: %w", err)
-			}
-			return nil
-		}
-	}
-
-	return errors.New("invalid TOTP code")
-}
-
-// DisableTOTP disables 2FA for a user
-func (s *DashboardAuthService) DisableTOTP(ctx context.Context, userID uuid.UUID, password string, ipAddress net.IP, userAgent string) error {
-	// Verify password
-	var passwordHash string
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT password_hash FROM platform.users WHERE id = $1 AND deleted_at IS NULL
-		`, userID).Scan(&passwordHash)
-	})
-	if err != nil {
-		return fmt.Errorf("failed to fetch user: %w", err)
-	}
-
-	err = bcrypt.CompareHashAndPassword([]byte(passwordHash), []byte(password))
-	if err != nil {
-		return errors.New("password is incorrect")
-	}
-
-	// Disable TOTP
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET totp_enabled = false, totp_secret = NULL, backup_codes = NULL, updated_at = NOW()
-			WHERE id = $1
-		`, userID)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to disable TOTP: %w", err)
-	}
-
-	// Log activity
-	s.logActivity(ctx, userID, "2fa_disable", "user", userID.String(), ipAddress, userAgent, nil)
-
-	return nil
-}
-
 // GetUserByID fetches a dashboard user by ID
 func (s *DashboardAuthService) GetUserByID(ctx context.Context, userID uuid.UUID) (*DashboardUser, error) {
 	user := &DashboardUser{}
@@ -798,422 +495,6 @@ func (s *DashboardAuthService) resolveTenantMembership(ctx context.Context, user
 		membership.tenantRole = tenantRole
 	}
 	return membership
-}
-
-// SSOIdentity represents a linked SSO identity for a dashboard user
-type SSOIdentity struct {
-	ID             uuid.UUID `json:"id"`
-	UserID         uuid.UUID `json:"user_id"`
-	Provider       string    `json:"provider"`
-	ProviderUserID string    `json:"provider_user_id"`
-	Email          *string   `json:"email,omitempty"`
-	CreatedAt      time.Time `json:"created_at"`
-}
-
-// GetUserBySSOIdentity finds a dashboard user by their SSO identity
-func (s *DashboardAuthService) GetUserBySSOIdentity(ctx context.Context, provider, providerUserID string) (*DashboardUser, error) {
-	// Split provider into type and name (format: "oauth:authelia" or "saml:okta")
-	parts := strings.SplitN(provider, ":", 2)
-	if len(parts) != 2 {
-		return nil, fmt.Errorf("invalid provider format (expected 'type:name'): %s", provider)
-	}
-	providerType := parts[0]
-	providerName := parts[1]
-
-	user := &DashboardUser{}
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT u.id, u.email, u.email_verified, u.full_name, u.avatar_url, u.role, u.totp_enabled,
-			       u.is_active, u.is_locked, u.last_login_at, u.created_at, u.updated_at
-			FROM platform.users u
-			INNER JOIN platform.sso_identities si ON si.user_id = u.id
-			WHERE si.provider_type = $1 AND si.provider_name = $2 AND si.provider_user_id = $3 AND u.deleted_at IS NULL
-		`, providerType, providerName, providerUserID).Scan(
-			&user.ID, &user.Email, &user.EmailVerified, &user.FullName, &user.AvatarURL,
-			&user.Role, &user.TOTPEnabled, &user.IsActive, &user.IsLocked, &user.LastLoginAt,
-			&user.CreatedAt, &user.UpdatedAt,
-		)
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, nil // No user found with this SSO identity
-		}
-		return nil, fmt.Errorf("failed to fetch user by SSO identity: %w", err)
-	}
-	return user, nil
-}
-
-// GetUserByEmail finds a dashboard user by email
-func (s *DashboardAuthService) GetUserByEmail(ctx context.Context, email string) (*DashboardUser, error) {
-	user := &DashboardUser{}
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT id, email, email_verified, full_name, avatar_url, role, totp_enabled,
-			       is_active, is_locked, last_login_at, created_at, updated_at
-			FROM platform.users
-			WHERE email = $1 AND deleted_at IS NULL
-		`, email).Scan(
-			&user.ID, &user.Email, &user.EmailVerified, &user.FullName, &user.AvatarURL,
-			&user.Role, &user.TOTPEnabled, &user.IsActive, &user.IsLocked, &user.LastLoginAt,
-			&user.CreatedAt, &user.UpdatedAt,
-		)
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, nil // No user found
-		}
-		return nil, fmt.Errorf("failed to fetch user by email: %w", err)
-	}
-	return user, nil
-}
-
-// FindOrCreateUserBySSO finds an existing user by SSO identity or email, or creates a new one
-// Returns the user and a boolean indicating if a new user was created
-func (s *DashboardAuthService) FindOrCreateUserBySSO(ctx context.Context, email, name, provider, providerUserID string) (*DashboardUser, bool, error) {
-	// First, try to find by SSO identity
-	user, err := s.GetUserBySSOIdentity(ctx, provider, providerUserID)
-	if err != nil {
-		return nil, false, err
-	}
-	if user != nil {
-		// Update last login
-		_ = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-			_, err := tx.Exec(ctx, `
-				UPDATE platform.users SET last_login_at = NOW() WHERE id = $1
-			`, user.ID)
-			return err
-		})
-		return user, false, nil
-	}
-
-	// Try to find by email
-	user, err = s.GetUserByEmail(ctx, email)
-	if err != nil {
-		return nil, false, err
-	}
-
-	if user != nil {
-		// Link SSO identity to existing user
-		err = s.LinkSSOIdentity(ctx, user.ID, provider, providerUserID, email)
-		if err != nil {
-			return nil, false, err
-		}
-		// Update last login
-		_ = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-			_, err := tx.Exec(ctx, `
-				UPDATE platform.users SET last_login_at = NOW() WHERE id = $1
-			`, user.ID)
-			return err
-		})
-		return user, false, nil
-	}
-
-	// Create new user (JIT provisioning)
-	// Split provider into type and name (format: "oauth:authelia" or "saml:okta")
-	parts := strings.SplitN(provider, ":", 2)
-	if len(parts) != 2 {
-		return nil, false, fmt.Errorf("invalid provider format (expected 'type:name'): %s", provider)
-	}
-	providerType := parts[0]
-	providerName := parts[1]
-
-	user = &DashboardUser{}
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		// Create user without password (SSO-only user)
-		err := tx.QueryRow(ctx, `
-			INSERT INTO platform.users (email, full_name, email_verified, is_active)
-			VALUES ($1, $2, true, true)
-			RETURNING id, email, email_verified, full_name, avatar_url, role, totp_enabled,
-			          is_active, is_locked, last_login_at, created_at, updated_at
-		`, email, name).Scan(
-			&user.ID, &user.Email, &user.EmailVerified, &user.FullName, &user.AvatarURL,
-			&user.Role, &user.TOTPEnabled, &user.IsActive, &user.IsLocked, &user.LastLoginAt,
-			&user.CreatedAt, &user.UpdatedAt,
-		)
-		if err != nil {
-			return err
-		}
-
-		// Link SSO identity
-		_, err = tx.Exec(ctx, `
-			INSERT INTO platform.sso_identities (user_id, provider_type, provider_name, provider_user_id, email)
-			VALUES ($1, $2, $3, $4, $5)
-		`, user.ID, providerType, providerName, providerUserID, email)
-		return err
-	})
-	if err != nil {
-		return nil, false, fmt.Errorf("failed to create user via SSO: %w", err)
-	}
-
-	return user, true, nil
-}
-
-// LinkSSOIdentity links an SSO identity to an existing dashboard user
-func (s *DashboardAuthService) LinkSSOIdentity(ctx context.Context, userID uuid.UUID, provider, providerUserID, email string) error {
-	// Split provider into type and name (format: "oauth:authelia" or "saml:okta")
-	parts := strings.SplitN(provider, ":", 2)
-	if len(parts) != 2 {
-		return fmt.Errorf("invalid provider format (expected 'type:name'): %s", provider)
-	}
-	providerType := parts[0]
-	providerName := parts[1]
-
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			INSERT INTO platform.sso_identities (user_id, provider_type, provider_name, provider_user_id, email)
-			VALUES ($1, $2, $3, $4, $5)
-			ON CONFLICT (provider_type, provider_name, provider_user_id) DO UPDATE SET email = EXCLUDED.email
-		`, userID, providerType, providerName, providerUserID, email)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to link SSO identity: %w", err)
-	}
-	return nil
-}
-
-// RequestPasswordReset creates a password reset token for a dashboard user
-// Returns the plaintext token (to be sent via email) or nil if user not found
-func (s *DashboardAuthService) RequestPasswordReset(ctx context.Context, email string) (string, error) {
-	// Find user by email
-	user, err := s.GetUserByEmail(ctx, email)
-	if err != nil {
-		return "", err
-	}
-	if user == nil {
-		// Don't reveal if user exists or not
-		return "", nil
-	}
-
-	// Generate a secure random token
-	tokenBytes := make([]byte, 32)
-	if _, err := rand.Read(tokenBytes); err != nil {
-		return "", fmt.Errorf("failed to generate token: %w", err)
-	}
-	token := base64.URLEncoding.EncodeToString(tokenBytes)
-
-	// Hash the token for storage
-	tokenHash := sha256.Sum256([]byte(token))
-	tokenHashHex := hex.EncodeToString(tokenHash[:])
-
-	// Delete any existing tokens for this user
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `DELETE FROM platform.password_reset_tokens WHERE user_id = $1`, user.ID)
-		return err
-	})
-	if err != nil {
-		return "", fmt.Errorf("failed to clean up old tokens: %w", err)
-	}
-
-	// Create new token (expires in 1 hour)
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `
-			INSERT INTO platform.password_reset_tokens (user_id, token, expires_at)
-			VALUES ($1, $2, NOW() + INTERVAL '1 hour')
-		`, user.ID, tokenHashHex)
-		return err
-	})
-	if err != nil {
-		return "", fmt.Errorf("failed to create password reset token: %w", err)
-	}
-
-	return token, nil
-}
-
-// VerifyPasswordResetToken verifies a password reset token is valid
-func (s *DashboardAuthService) VerifyPasswordResetToken(ctx context.Context, token string) (bool, error) {
-	// Hash the token for lookup
-	tokenHash := sha256.Sum256([]byte(token))
-	tokenHashHex := hex.EncodeToString(tokenHash[:])
-
-	var exists bool
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT EXISTS(
-				SELECT 1 FROM platform.password_reset_tokens
-				WHERE token = $1 AND expires_at > NOW() AND used = false
-			)
-		`, tokenHashHex).Scan(&exists)
-	})
-	if err != nil {
-		return false, fmt.Errorf("failed to verify token: %w", err)
-	}
-
-	return exists, nil
-}
-
-// ResetPassword resets a dashboard user's password using a valid reset token
-func (s *DashboardAuthService) ResetPassword(ctx context.Context, token, newPassword string) error {
-	// Validate new password length
-	if len(newPassword) < MinPasswordLength {
-		return fmt.Errorf("password must be at least %d characters", MinPasswordLength)
-	}
-	if len(newPassword) > MaxPasswordLength {
-		return fmt.Errorf("password must be at most %d characters", MaxPasswordLength)
-	}
-
-	// Hash the token for lookup
-	tokenHash := sha256.Sum256([]byte(token))
-	tokenHashHex := hex.EncodeToString(tokenHash[:])
-
-	// Find the token and get user ID
-	var userID uuid.UUID
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, `
-			SELECT user_id FROM platform.password_reset_tokens
-			WHERE token = $1 AND expires_at > NOW() AND used = false
-		`, tokenHashHex).Scan(&userID)
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return errors.New("invalid or expired password reset token")
-		}
-		return fmt.Errorf("failed to verify token: %w", err)
-	}
-
-	// Hash new password
-	newHash, err := bcrypt.GenerateFromPassword([]byte(newPassword), DefaultBcryptCost)
-	if err != nil {
-		return fmt.Errorf("failed to hash password: %w", err)
-	}
-
-	// Update password and mark token as used
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		// Update password
-		_, err := tx.Exec(ctx, `
-			UPDATE platform.users
-			SET password_hash = $1, updated_at = NOW()
-			WHERE id = $2
-		`, newHash, userID)
-		if err != nil {
-			return err
-		}
-
-		// Mark token as used
-		_, err = tx.Exec(ctx, `
-			UPDATE platform.password_reset_tokens
-			SET used = true, used_at = NOW()
-			WHERE token = $1
-		`, tokenHashHex)
-		return err
-	})
-	if err != nil {
-		return fmt.Errorf("failed to reset password: %w", err)
-	}
-
-	return nil
-}
-
-// LoginViaSSO logs in a dashboard user via SSO and returns tokens
-func (s *DashboardAuthService) LoginViaSSO(ctx context.Context, user *DashboardUser, ipAddress net.IP, userAgent string) (*LoginResponse, error) {
-	// Safe IP address string for logging
-	var ipStr string
-	if ipAddress != nil {
-		ipStr = ipAddress.String()
-	}
-
-	// Check if account is locked
-	if user.IsLocked {
-		LogSecurityWarning(ctx, SecurityEvent{
-			Type:      SecurityEventLoginFailed,
-			UserID:    user.ID.String(),
-			Email:     user.Email,
-			IPAddress: ipStr,
-			UserAgent: userAgent,
-			Details:   map[string]interface{}{"reason": "account_locked", "dashboard": true, "sso": true},
-		})
-		return nil, ErrAccountLocked
-	}
-
-	// Check if account is active
-	if !user.IsActive {
-		LogSecurityWarning(ctx, SecurityEvent{
-			Type:      SecurityEventLoginFailed,
-			UserID:    user.ID.String(),
-			Email:     user.Email,
-			IPAddress: ipStr,
-			UserAgent: userAgent,
-			Details:   map[string]interface{}{"reason": "account_inactive", "dashboard": true, "sso": true},
-		})
-		return nil, errors.New("account is inactive")
-	}
-
-	// Log successful SSO login
-	LogSecurityEvent(ctx, SecurityEvent{
-		Type:      SecurityEventLoginSuccess,
-		UserID:    user.ID.String(),
-		Email:     user.Email,
-		IPAddress: ipStr,
-		UserAgent: userAgent,
-		Details:   map[string]interface{}{"dashboard": true, "sso": true},
-	})
-
-	// Prepare user metadata for JWT
-	userMetadata := map[string]interface{}{}
-	if user.FullName != nil {
-		userMetadata["name"] = *user.FullName
-	}
-	if user.AvatarURL != nil {
-		userMetadata["avatar"] = *user.AvatarURL
-	}
-
-	// Use the actual role from the database, defaulting to dashboard_user if empty
-	userRole := user.Role
-	if userRole == "" {
-		userRole = "dashboard_user"
-	}
-
-	// Determine tenant context for JWT claims
-	tenantOpts := TenantTokenOptions{
-		IsInstanceAdmin: userRole == "instance_admin",
-	}
-
-	membership := s.resolveTenantMembership(ctx, user.ID)
-	if membership.tenantID != nil {
-		tenantOpts.TenantID = membership.tenantID
-		tenantOpts.TenantRole = membership.tenantRole
-	}
-
-	// Generate JWT token pair with tenant context
-	accessToken, refreshToken, sessionID, err := s.jwtManager.GenerateTokenPairWithTenant(user.ID.String(), user.Email, userRole, userMetadata, nil, tenantOpts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate tokens: %w", err)
-	}
-
-	// Hash the access token
-	hash := sha256.Sum256([]byte(accessToken))
-	tokenHash := hex.EncodeToString(hash[:])
-
-	// Handle nil IP address
-	var ipAddressStr interface{}
-	if ipAddress != nil {
-		ipAddressStr = ipAddress.String()
-	}
-
-	// Delete existing sessions and create new one
-	err = database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, `DELETE FROM platform.sessions WHERE user_id = $1`, user.ID)
-		if err != nil {
-			return err
-		}
-		_, err = tx.Exec(ctx, `
-			INSERT INTO platform.sessions (id, user_id, token, ip_address, user_agent, expires_at)
-			VALUES ($1, $2, $3, $4, $5, NOW() + INTERVAL '24 hours')
-		`, sessionID, user.ID, tokenHash, ipAddressStr, userAgent)
-		return err
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create session: %w", err)
-	}
-
-	// Log activity
-	s.logActivity(ctx, user.ID, "sso_login", "", "", ipAddress, userAgent, nil)
-
-	return &LoginResponse{
-		AccessToken:  accessToken,
-		RefreshToken: refreshToken,
-		ExpiresIn:    int64(24 * 60 * 60),
-	}, nil
 }
 
 // RefreshToken generates a new access token using a refresh token for dashboard users
